Minor changes to both BSTNode and BSTree
Added nodiscard attribute to methods that should have it.

Changed the order the copy constructor and assignment operator use to copy the data from one tree to other. Now they use width order which guaranties that the new tree is an exact copy of the previous one and that the nodes don´t get disordered.

Deleted a ternary operator from the add method of the BSTree class and changed it with a conventional if else, otherwise I had to cast the first expression and the simplicity of the ternary operator was lost